Cleanup allocations in routing
