--- .idea/runConfigurations/Rebuild__JMH.xml
@@ -1,21 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Rebuild  JMH" type="GradleRunConfiguration" factoryName="Gradle">
-    <ExternalSystemSettings>
-      <option name="executionName" />
-      <option name="externalProjectPath" value="$PROJECT_DIR$/ktor-server/ktor-server-benchmarks" />
-      <option name="externalSystemIdString" value="GRADLE" />
-      <option name="scriptParameters" value="" />
-      <option name="taskDescriptions">
-        <list />
-      </option>
-      <option name="taskNames">
-        <list>
-          <option value="jmhJar" />
-        </list>
-      </option>
-      <option name="vmOptions" value="" />
-    </ExternalSystemSettings>
-    <extension name="coverage" enabled="false" merge="false" sample_coverage="true" runner="idea" />
-    <method />
-  </configuration>
-</component>
