+++ ktor-server/ktor-server-core/jvmAndNix/test/io/ktor/server/application/RouteScopedPluginTest.kt
@@ -10,9 +10,7 @@ import io.ktor.server.response.*
 import io.ktor.server.routing.*
 import io.ktor.server.testing.*
 import io.ktor.util.*
-import io.ktor.util.collections.*
 import io.ktor.util.pipeline.*
-import io.ktor.utils.io.concurrent.*
 import kotlin.test.*
 
 @Suppress("DEPRECATION")
