+++ ktor-server/ktor-server-host-common/jvm/src/io/ktor/server/engine/BaseApplicationResponseJvm.kt
@@ -44,7 +44,7 @@ public actual abstract class BaseApplicationResponse actual constructor(
     public actual final override val pipeline: ApplicationSendPipeline = ApplicationSendPipeline(
         call.application.environment.developmentMode
     ).apply {
-        merge(call.application.sendPipeline)
+        resetFrom(call.application.sendPipeline)
     }
 
     /**
