+++ ktor-server/ktor-server-plugins/ktor-server-default-headers/api/ktor-server-default-headers.api
@@ -1,8 +1,12 @@
 public final class io/ktor/server/plugins/defaultheaders/DefaultHeadersConfig {
 	public fun <init> ()V
-	public final fun getClock ()Lkotlin/jvm/functions/Function0;
+	public final fun getClock ()Lio/ktor/server/plugins/defaultheaders/DefaultHeadersConfig$Clock;
 	public final fun header (Ljava/lang/String;Ljava/lang/String;)V
-	public final fun setClock (Lkotlin/jvm/functions/Function0;)V
+	public final fun setClock (Lio/ktor/server/plugins/defaultheaders/DefaultHeadersConfig$Clock;)V
+}
+
+public abstract interface class io/ktor/server/plugins/defaultheaders/DefaultHeadersConfig$Clock {
+	public abstract fun now ()J
 }
 
 public final class io/ktor/server/plugins/defaultheaders/DefaultHeadersKt {
