+++ ktor-server/ktor-server-tests/jvm/test/io/ktor/server/plugins/CompressionTest.kt
@@ -17,7 +17,6 @@ import io.ktor.server.routing.*
 import io.ktor.server.testing.*
 import io.ktor.utils.io.*
 import kotlinx.coroutines.*
-import java.io.*
 import java.time.*
 import java.util.zip.*
 import kotlin.coroutines.*
