+++ ktor-server/ktor-server-tests/jvm/test/io/ktor/server/plugins/PartialContentTest.kt
@@ -7,7 +7,6 @@ package io.ktor.server.plugins
 import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
-import io.ktor.server.http.content.*
 import io.ktor.server.plugins.autohead.*
 import io.ktor.server.plugins.cachingheaders.*
 import io.ktor.server.plugins.conditionalheaders.*
