+++ ktor-server/ktor-server-tests/jvm/test/io/ktor/server/sessions/SessionTest.kt
@@ -13,7 +13,6 @@ import io.ktor.server.routing.*
 import io.ktor.server.testing.*
 import io.ktor.util.*
 import io.ktor.util.date.*
-import io.ktor.utils.io.jvm.javaio.*
 import kotlinx.coroutines.*
 import kotlin.random.*
 import kotlin.test.*
