+++ ktor-server/ktor-server-tests/jvmAndNix/test/io/ktor/tests/server/routing/RoutingProcessingTest.kt
@@ -718,307 +718,24 @@ class RoutingProcessingTest {
     }
 
     @Test
-    fun testRoutingWithTracing() = withTestApplication {
-        var trace: RoutingResolveTrace? = null
-        application.routing {
-            trace {
-                trace = it
-            }
-            get("/bar") { call.respond("/bar") }
-            get("/baz") { call.respond("/baz") }
-            get("/baz/x") { call.respond("/baz/x") }
-            get("/baz/x/{optional?}") { call.respond("/baz/x/{optional?}") }
-            get("/baz/{y}") { call.respond("/baz/{y}") }
-            get("/baz/{y}/value") { call.respond("/baz/{y}/value") }
-            get("/{param}") { call.respond("/{param}") }
-            get("/{param}/x") { call.respond("/{param}/x") }
-            get("/{param}/x/z") { call.respond("/{param}/x/z") }
-            get("/*/extra") { call.respond("/*/extra") }
-            header("a", "x") { get { call.respond("a") } }
-            header("b", "x") { get { call.respond("b") } }
-        }
-
-        handleRequest {
-            uri = "/bar"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/bar", it.response.content)
-            assertEquals(
-                """Trace for [bar]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:1 -> SUCCESS @ /bar
-    /bar/(method:GET), segment:1 -> SUCCESS @ /bar/(method:GET)
-  /baz, segment:0 -> FAILURE "Selector didn't match" @ /baz
-  /{param}, segment:0 -> FAILURE "Better match was already found" @ /{param}
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "bar" -> "(method:GET)"
-Route resolve result:
-  SUCCESS @ /bar/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/bar/x"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/{param}/x", it.response.content)
-            assertEquals(
-                """Trace for [bar, x]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:1 -> SUCCESS @ /bar
-    /bar/(method:GET), segment:1 -> SUCCESS @ /bar/(method:GET)
-      /bar/(method:GET), segment:1 -> FAILURE "Not all segments matched" @ /bar/(method:GET)
-  /baz, segment:0 -> FAILURE "Selector didn't match" @ /baz
-  /{param}, segment:1 -> SUCCESS; Parameters [param=[bar]] @ /{param}
-    /{param}/(method:GET), segment:1 -> SUCCESS @ /{param}/(method:GET)
-      /{param}/(method:GET), segment:1 -> FAILURE "Not all segments matched" @ /{param}/(method:GET)
-    /{param}/x, segment:2 -> SUCCESS @ /{param}/x
-      /{param}/x/(method:GET), segment:2 -> SUCCESS @ /{param}/x/(method:GET)
-      /{param}/x/z, segment:2 -> FAILURE "Selector didn't match" @ /{param}/x/z
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "{param}" -> "x" -> "(method:GET)"
-Route resolve result:
-  SUCCESS; Parameters [param=[bar]] @ /{param}/x/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/baz/x"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/baz/x", it.response.content)
-            assertEquals(
-                """Trace for [baz, x]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:0 -> FAILURE "Selector didn't match" @ /bar
-  /baz, segment:1 -> SUCCESS @ /baz
-    /baz/(method:GET), segment:1 -> SUCCESS @ /baz/(method:GET)
-      /baz/(method:GET), segment:1 -> FAILURE "Not all segments matched" @ /baz/(method:GET)
-    /baz/x, segment:2 -> SUCCESS @ /baz/x
-      /baz/x/(method:GET), segment:2 -> SUCCESS @ /baz/x/(method:GET)
-      /baz/x/{optional?}, segment:2 -> FAILURE "Better match was already found" @ /baz/x/{optional?}
-    /baz/{y}, segment:1 -> FAILURE "Better match was already found" @ /baz/{y}
-  /{param}, segment:0 -> FAILURE "Better match was already found" @ /{param}
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "baz" -> "x" -> "(method:GET)"
-Route resolve result:
-  SUCCESS @ /baz/x/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/baz/doo"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/baz/{y}", it.response.content)
-            assertEquals(
-                """Trace for [baz, doo]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:0 -> FAILURE "Selector didn't match" @ /bar
-  /baz, segment:1 -> SUCCESS @ /baz
-    /baz/(method:GET), segment:1 -> SUCCESS @ /baz/(method:GET)
-      /baz/(method:GET), segment:1 -> FAILURE "Not all segments matched" @ /baz/(method:GET)
-    /baz/x, segment:1 -> FAILURE "Selector didn't match" @ /baz/x
-    /baz/{y}, segment:2 -> SUCCESS; Parameters [y=[doo]] @ /baz/{y}
-      /baz/{y}/(method:GET), segment:2 -> SUCCESS @ /baz/{y}/(method:GET)
-      /baz/{y}/value, segment:2 -> FAILURE "Selector didn't match" @ /baz/{y}/value
-  /{param}, segment:0 -> FAILURE "Better match was already found" @ /{param}
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "baz" -> "{y}" -> "(method:GET)"
-Route resolve result:
-  SUCCESS; Parameters [y=[doo]] @ /baz/{y}/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/baz/x/z"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/baz/x/{optional?}", it.response.content)
-            assertEquals(
-                """Trace for [baz, x, z]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:0 -> FAILURE "Selector didn't match" @ /bar
-  /baz, segment:1 -> SUCCESS @ /baz
-    /baz/(method:GET), segment:1 -> SUCCESS @ /baz/(method:GET)
-      /baz/(method:GET), segment:1 -> FAILURE "Not all segments matched" @ /baz/(method:GET)
-    /baz/x, segment:2 -> SUCCESS @ /baz/x
-      /baz/x/(method:GET), segment:2 -> SUCCESS @ /baz/x/(method:GET)
-        /baz/x/(method:GET), segment:2 -> FAILURE "Not all segments matched" @ /baz/x/(method:GET)
-      /baz/x/{optional?}, segment:3 -> SUCCESS; Parameters [optional=[z]] @ /baz/x/{optional?}
-        /baz/x/{optional?}/(method:GET), segment:3 -> SUCCESS @ /baz/x/{optional?}/(method:GET)
-    /baz/{y}, segment:1 -> FAILURE "Better match was already found" @ /baz/{y}
-  /{param}, segment:0 -> FAILURE "Better match was already found" @ /{param}
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "baz" -> "x" -> "{optional?}" -> "(method:GET)"
-Route resolve result:
-  SUCCESS; Parameters [optional=[z]] @ /baz/x/{optional?}/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/baz/x/value"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/baz/x/{optional?}", it.response.content)
-            assertEquals(
-                """Trace for [baz, x, value]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:0 -> FAILURE "Selector didn't match" @ /bar
-  /baz, segment:1 -> SUCCESS @ /baz
-    /baz/(method:GET), segment:1 -> SUCCESS @ /baz/(method:GET)
-      /baz/(method:GET), segment:1 -> FAILURE "Not all segments matched" @ /baz/(method:GET)
-    /baz/x, segment:2 -> SUCCESS @ /baz/x
-      /baz/x/(method:GET), segment:2 -> SUCCESS @ /baz/x/(method:GET)
-        /baz/x/(method:GET), segment:2 -> FAILURE "Not all segments matched" @ /baz/x/(method:GET)
-      /baz/x/{optional?}, segment:3 -> SUCCESS; Parameters [optional=[value]] @ /baz/x/{optional?}
-        /baz/x/{optional?}/(method:GET), segment:3 -> SUCCESS @ /baz/x/{optional?}/(method:GET)
-    /baz/{y}, segment:1 -> FAILURE "Better match was already found" @ /baz/{y}
-  /{param}, segment:0 -> FAILURE "Better match was already found" @ /{param}
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "baz" -> "x" -> "{optional?}" -> "(method:GET)"
-Route resolve result:
-  SUCCESS; Parameters [optional=[value]] @ /baz/x/{optional?}/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/p"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/{param}", it.response.content)
-            assertEquals(
-                """Trace for [p]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:0 -> FAILURE "Selector didn't match" @ /bar
-  /baz, segment:0 -> FAILURE "Selector didn't match" @ /baz
-  /{param}, segment:1 -> SUCCESS; Parameters [param=[p]] @ /{param}
-    /{param}/(method:GET), segment:1 -> SUCCESS @ /{param}/(method:GET)
-    /{param}/x, segment:1 -> FAILURE "Selector didn't match" @ /{param}/x
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "{param}" -> "(method:GET)"
-Route resolve result:
-  SUCCESS; Parameters [param=[p]] @ /{param}/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/p/x"
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("/{param}/x", it.response.content)
-            assertEquals(
-                """Trace for [p, x]
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:0 -> FAILURE "Selector didn't match" @ /bar
-  /baz, segment:0 -> FAILURE "Selector didn't match" @ /baz
-  /{param}, segment:1 -> SUCCESS; Parameters [param=[p]] @ /{param}
-    /{param}/(method:GET), segment:1 -> SUCCESS @ /{param}/(method:GET)
-      /{param}/(method:GET), segment:1 -> FAILURE "Not all segments matched" @ /{param}/(method:GET)
-    /{param}/x, segment:2 -> SUCCESS @ /{param}/x
-      /{param}/x/(method:GET), segment:2 -> SUCCESS @ /{param}/x/(method:GET)
-      /{param}/x/z, segment:2 -> FAILURE "Selector didn't match" @ /{param}/x/z
-  /*, segment:0 -> FAILURE "Better match was already found" @ /*
-  /(header:a = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:a = x)
-  /(header:b = x), segment:0 -> FAILURE "Selector didn't match" @ /(header:b = x)
-Matched routes:
-  "" -> "{param}" -> "x" -> "(method:GET)"
-Route resolve result:
-  SUCCESS; Parameters [param=[p]] @ /{param}/x/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-
-        handleRequest {
-            uri = "/"
-            addHeader("a", "x")
-            addHeader("b", "x")
-            method = HttpMethod.Get
-        }.let {
-            assertEquals("a", it.response.content)
-            assertEquals(
-                """Trace for []
-/, segment:0 -> SUCCESS @ /
-  /bar, segment:0 -> FAILURE "Selector didn't match" @ /bar
-  /baz, segment:0 -> FAILURE "Selector didn't match" @ /baz
-  /{param}, segment:0 -> FAILURE "Selector didn't match" @ /{param}
-  /*, segment:0 -> FAILURE "Selector didn't match" @ /*
-  /(header:a = x), segment:0 -> SUCCESS @ /(header:a = x)
-    /(header:a = x)/(method:GET), segment:0 -> SUCCESS @ /(header:a = x)/(method:GET)
-  /(header:b = x), segment:0 -> SUCCESS @ /(header:b = x)
-    /(header:b = x)/(method:GET), segment:0 -> SUCCESS @ /(header:b = x)/(method:GET)
-Matched routes:
-  "" -> "(header:a = x)" -> "(method:GET)"
-  "" -> "(header:b = x)" -> "(method:GET)"
-Route resolve result:
-  SUCCESS @ /(header:a = x)/(method:GET)
-""",
-                trace?.buildText()
-            )
-        }
-    }
-
-    @Test
-    fun testRouteWithParameterPrefixAndSuffixHasMorePriority() = withTestApplication {
-        application.routing {
-            get("/foo:{baz}") {
-                call.respondText("foo")
-            }
-            get("/{baz}") {
-                call.respondText("baz")
-            }
-            get("/{baz}:bar") {
-                call.respondText("bar")
+    fun testRouteWithParameterPrefixAndSuffixHasMorePriority() = testApplication {
+        application {
+            routing {
+                get("/foo:{baz}") {
+                    call.respondText("foo")
+                }
+                get("/{baz}") {
+                    call.respondText("baz")
+                }
+                get("/{baz}:bar") {
+                    call.respondText("bar")
+                }
             }
         }
 
-        handleRequest(HttpMethod.Get, "/foo:bar").let { call ->
-            assertEquals(call.response.content, "foo")
-        }
-
-        handleRequest(HttpMethod.Get, "/baz").let { call ->
-            assertEquals(call.response.content, "baz")
-        }
-
-        handleRequest(HttpMethod.Get, "/baz:bar").let { call ->
-            assertEquals(call.response.content, "bar")
-        }
+        assertEquals("foo", client.get("/foo:bar").bodyAsText())
+        assertEquals("baz", client.get("/baz").bodyAsText())
+        assertEquals("bar", client.get("/baz:bar").bodyAsText())
     }
 
     @Test
